requirements.txt
streamlit>=1.32
pandas>=2.0
openpyxl>=3.1
pydantic>=2.0
pyyaml>=6.0
fpdf2>=2.7